feat(map): add NavigationButtons component
